SSCS-9333: Move related fields to separate class
