+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/domain/SscsFinalDecisionCaseData.java
@@ -1,0 +1,51 @@
+package uk.gov.hmcts.reform.sscs.ccd.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
+import uk.gov.hmcts.reform.sscs.ccd.validation.documentlink.DocumentLinkMustBePdf;
+import uk.gov.hmcts.reform.sscs.ccd.validation.groups.UniversalCreditValidationGroup;
+import uk.gov.hmcts.reform.sscs.ccd.validation.localdate.LocalDateMustNotBeInFuture;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@Builder(toBuilder = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class SscsFinalDecisionCaseData {
+
+    private String writeFinalDecisionIsDescriptorFlow;
+    private String writeFinalDecisionGenerateNotice;
+    private String writeFinalDecisionAllowedOrRefused;
+    private String writeFinalDecisionTypeOfHearing;
+    private String writeFinalDecisionPresentingOfficerAttendedQuestion;
+    private String writeFinalDecisionAppellantAttendedQuestion;
+    private String writeFinalDecisionStartDate;
+    private String writeFinalDecisionEndDateType;
+    private String writeFinalDecisionEndDate;
+    private String writeFinalDecisionDisabilityQualifiedPanelMemberName;
+    private String writeFinalDecisionMedicallyQualifiedPanelMemberName;
+    private String writeFinalDecisionOtherPanelMemberName;
+    @LocalDateMustNotBeInFuture(message = "Decision notice date of decision must not be in the future", groups = UniversalCreditValidationGroup.class)
+    private String writeFinalDecisionDateOfDecision;
+    private String writeFinalDecisionDetailsOfDecision;
+    private List<CollectionItem<String>> writeFinalDecisionReasons;
+    private String writeFinalDecisionPageSectionReference;
+    private String writeFinalDecisionAnythingElse;
+    @DocumentLinkMustBePdf(message = "You need to upload PDF documents only", groups = UniversalCreditValidationGroup.class)
+    private DocumentLink writeFinalDecisionPreviewDocument;
+    private String writeFinalDecisionGeneratedDate;
+
+    @JsonIgnore
+    public boolean isDailyLivingAndOrMobilityDecision() {
+        return StringUtils.equalsIgnoreCase("yes", writeFinalDecisionIsDescriptorFlow);
+    }
+
+}
